+++ changelog.md
@@ -1,5 +1,7 @@
 ##### Version 1.21.1-12.4.2-pre
 - Add feature for the item router to filter by mod in addition to tags (BluSunrize)
+- Change extracting conveyors to have configurable stack-size instead of configurable tickrate (BluSunrize)
+    - This requires existing conveyors to be adjusted again
 - Fix mixer accidentally crafting a lingering awkward potion (BluSunrize)
 - Fix catwalks not working properly when using a hammer (BluSunrize)
 - Fix shift-click behavior in the Engineer's Crafting Table (BluSunrize)